moved ViewRenderer to its own file
